+++ Moco/Moco/MocoBounds.cpp
@@ -20,6 +20,8 @@
 
 #include "MocoUtilities.h"
 
+#include <OpenSim/Common/Constant.h>
+
 using namespace OpenSim;
 
 MocoBounds::MocoBounds() {
@@ -33,7 +35,7 @@ MocoBounds::MocoBounds(double value) : MocoBounds() {
 }
 
 MocoBounds::MocoBounds(double lower, double upper) : MocoBounds() {
-    OPENSIM_THROW_IF(SimTK::isNaN(lower) || SimTK::isNaN(upper), Exception, 
+    OPENSIM_THROW_IF(SimTK::isNaN(lower) || SimTK::isNaN(upper), Exception,
         "NaN value detected. Please provide a non-NaN values for the bounds.");
     OPENSIM_THROW_IF(lower > upper, Exception,
         format("Expected lower <= upper, but lower=%g and upper=%g.",
@@ -68,4 +70,36 @@ void MocoBounds::printDescription(std::ostream& stream) const {
         stream << "[" << getLower() << ", " << getUpper() << "]";
     }
     stream.flush();
+}
+
+MocoFunctionBounds::MocoFunctionBounds() {
+    constructProperties();
+}
+
+MocoFunctionBounds::MocoFunctionBounds(const Function& lower)
+        : MocoFunctionBounds() {
+    set_lower_bound(lower);
+}
+
+MocoFunctionBounds::MocoFunctionBounds(const Function& lower,
+        const Function& upper) : MocoFunctionBounds(lower) {
+    set_upper_bound(upper);
+}
+
+
+void MocoFunctionBounds::printDescription(std::ostream& stream) const {
+    if (isEquality()) {
+        stream << get_lower_bound().getConcreteClassName();
+    } else {
+        const auto& lower = get_lower_bound().getConcreteClassName();
+        const auto& upper = get_upper_bound().getConcreteClassName();
+        stream << "[" << lower << ", " << upper << "]";
+    }
+    stream.flush();
+}
+
+void MocoFunctionBounds::constructProperties() {
+    constructProperty_lower_bound(Constant(0));
+    constructProperty_upper_bound();
+    constructProperty_equality_with_lower(false);
 }