+++ Moco/Moco/MocoUtilities.h
@@ -565,6 +565,23 @@ void checkPropertyIsPositive(const Object& obj, const Property<T>& p) {
     }
 }
 
+/// Throw an exception if the property's value is negative.
+/// We assume that `p` is a single-value property.
+template <typename T>
+void checkPropertyIsNonnegative(const Object& obj, const Property<T>& p) {
+    const auto& value = p.getValue();
+    if (value < 0) {
+        std::stringstream msg;
+        msg << "Property '" << p.getName() << "' (in ";
+        if (!obj.getName().empty()) {
+            msg << "object '" << obj.getName() << "' of type ";
+        }
+        msg << obj.getConcreteClassName() << ") must be non-negative, but is "
+                << value << ".";
+        OPENSIM_THROW(Exception, msg.str());
+    }
+}
+
 /// Throw an exception if the property's value is neither in the provided
 /// range nor in the provided set.
 /// We assume that `p` is a single-value property.
@@ -771,6 +788,12 @@ TimeSeriesTable createExternalLoadsTableForGait(Model model,
         const MocoTrajectory& trajectory,
         const std::vector<std::string>& forceNamesRightFoot,
         const std::vector<std::string>& forceNamesLeftFoot);
+
+/// TODO
+OSIMMOCO_API
+void setKinematicStateFunctionBoundsFromTable(const Model& model,
+        const TimeSeriesTable& kinematics, const double& rangeRotational,
+        const double& rangeTranslational, MocoProblem& problem);
 
 } // namespace OpenSim